Make ScopeConfiguration publicly accessible
When using a hierarchy of Spring contexts, custom scopes are not visible in child contexts. With this change, it is now possible for users to import the configuration class into child contexts and use job and step scope within them.
